Added flag in install_deps.sh to allow running autotools.

+++ install_deps.sh
@@ -11,15 +11,21 @@ mkdir -p ${LOCAL_DEPS_DIR}/include
 # searches for libgflags.
 export LD_LIBRARY_PATH="${LD_LIBRARY_PATH}:${LOCAL_DEPS_DIR}/lib:"
 
+# Set to true to run autoconf and automake (sometimes necessary in Mac OS-X).
+RUN_AUTOTOOLS=false
+
 # Install gflags.
 echo ""
 echo "Installing gflags..."
 tar -zxf gflags-2.0-no-svn-files.tar.gz
 cd gflags-2.0
-rm missing
-aclocal
-autoconf
-automake --add-missing
+if ${RUN_AUTOTOOLS}
+then
+    rm missing
+    aclocal
+    autoconf
+    automake --add-missing
+fi
 ./configure --prefix=${LOCAL_DEPS_DIR} && make && make install
 cd ..
 echo "Done."
@@ -29,10 +35,13 @@ echo ""
 echo "Installing glog..."
 tar -zxf glog-0.3.2.tar.gz
 cd glog-0.3.2
-rm missing
-aclocal
-autoconf
-automake --add-missing
+if ${RUN_AUTOTOOLS}
+then
+    rm missing
+    aclocal
+    autoconf
+    automake --add-missing
+fi
 ./configure --prefix=${LOCAL_DEPS_DIR} --with-gflags=${LOCAL_DEPS_DIR} \
   && make && make install
 cd ..
